ai_1/requirements.txt
# CS4241 manual RAG stack (no LangChain) — pin versions in your viva README if needed.
pandas>=2.0.0
pymupdf>=1.24.0
numpy>=1.24.0
faiss-cpu>=1.8.0
sentence-transformers>=3.0.0
torch>=2.0.0
streamlit>=1.28.0
